add base class for MultiTemporal data readers
