+++ ascat/level1.py
@@ -40,6 +40,7 @@ import ascat.read_native.eps_native as eps_native
 import ascat.read_native.bufr as bufr
 import ascat.read_native.nc as nc
 import ascat.read_native.hdf5 as h5
+from ascat.base import ASCAT_MultiTemporalImageBase
 
 
 byte_nan = np.iinfo(np.byte).min
@@ -111,7 +112,7 @@ class AscatL1Image(ImageBase):
     def close(self):
         pass
 
-class AscatL1Bufr(MultiTemporalImageBase):
+class AscatL1Bufr(ASCAT_MultiTemporalImageBase):
     """
     Class for reading multiple ASCAT level1 images in bufr format.
 
@@ -191,44 +192,8 @@ class AscatL1Bufr(MultiTemporalImageBase):
             return datetime.strptime(orbit_start_str,
                                      self.filename_datetime_format[2])
 
-    def tstamps_for_daterange(self, startdate, enddate):
-        """
-        Returns the possible timestamps of the given daterange as a datetime
-        array.
 
-        Parameters
-        ----------
-        start_date : datetime.date or datetime.datetime
-            start date
-        end_date : datetime.date or datetime.datetime
-            end date
-
-        Returns
-        -------
-        dates : list
-            list of datetimes
-        """
-        file_list = []
-        delta_all = enddate - startdate
-        timestamps = []
-
-        for i in range(delta_all.days + 1):
-            timestamp = startdate + timedelta(days=i)
-
-            files = self._search_files(
-                timestamp, custom_templ=self.day_search_str)
-
-            file_list.extend(sorted(files))
-
-        for filename in file_list:
-            timestamps.append(self._get_orbit_start_date(filename))
-
-        timestamps = [dt for dt in timestamps if (
-                dt >= startdate and dt <= enddate)]
-        return timestamps
-
-
-class AscatL1Eps(MultiTemporalImageBase):
+class AscatL1Eps(ASCAT_MultiTemporalImageBase):
     """
     Class for reading multiple ASCAT level1 images in eps format.
 
@@ -276,71 +241,8 @@ class AscatL1Eps(MultiTemporalImageBase):
                                              datetime_format=datetime_format,
                                              exact_templ=False)
 
-    def _get_orbit_start_date(self, filename):
-        """
-        Returns the datetime of the file.
 
-        Parameters
-        ----------
-        filename : full name (including the path) of the file
-
-        Returns
-        -------
-        dates : datetime object
-            datetime from the filename
-        """
-        # if your data comes from the EUMETSAT EO Portal this function can
-        if self.eo_portal == True:
-            filename_base = os.path.basename(filename)
-            fln_spl = filename_base.split('_')[4]
-            fln_datetime = fln_spl[:-1]
-            return datetime.strptime(fln_datetime, self.datetime_format)
-
-        else:
-            orbit_start_str = os.path.basename(filename)[
-                              self.filename_datetime_format[0]:
-                              self.filename_datetime_format[1]]
-            return datetime.strptime(orbit_start_str,
-                                     self.filename_datetime_format[2])
-
-    def tstamps_for_daterange(self, startdate, enddate):
-        """
-        Returns the possible timestamps of the given daterange as a datetime
-        array.
-
-        Parameters
-        ----------
-        start_date : datetime.date or datetime.datetime
-            start date
-        end_date : datetime.date or datetime.datetime
-            end date
-
-        Returns
-        -------
-        dates : list
-            list of datetimes
-        """
-        file_list = []
-        delta_all = enddate - startdate
-        timestamps = []
-
-        for i in range(delta_all.days + 1):
-            timestamp = startdate + timedelta(days=i)
-
-            files = self._search_files(
-                timestamp, custom_templ=self.day_search_str)
-
-            file_list.extend(sorted(files))
-
-        for filename in file_list:
-            timestamps.append(self._get_orbit_start_date(filename))
-
-        timestamps = [dt for dt in timestamps if (
-                dt >= startdate and dt <= enddate)]
-        return timestamps
-
-
-class AscatL1Hdf5(MultiTemporalImageBase):
+class AscatL1Hdf5(ASCAT_MultiTemporalImageBase):
     """
     Class for reading multiple ASCAT level1 images in hdf5 format.
 
@@ -388,71 +290,8 @@ class AscatL1Hdf5(MultiTemporalImageBase):
                                              datetime_format=datetime_format,
                                              exact_templ=False)
 
-    def _get_orbit_start_date(self, filename):
-        """
-        Returns the datetime of the file.
 
-        Parameters
-        ----------
-        filename : full name (including the path) of the file
-
-        Returns
-        -------
-        dates : datetime object
-            datetime from the filename
-        """
-        # if your data comes from the EUMETSAT EO Portal this function can
-        if self.eo_portal == True:
-            filename_base = os.path.basename(filename)
-            fln_spl = filename_base.split('_')[4]
-            fln_datetime = fln_spl[:-1]
-            return datetime.strptime(fln_datetime, self.datetime_format)
-
-        else:
-            orbit_start_str = os.path.basename(filename)[
-                              self.filename_datetime_format[0]:
-                              self.filename_datetime_format[1]]
-            return datetime.strptime(orbit_start_str,
-                                     self.filename_datetime_format[2])
-
-    def tstamps_for_daterange(self, startdate, enddate):
-        """
-        Returns the possible timestamps of the given daterange as a datetime
-        array.
-
-        Parameters
-        ----------
-        start_date : datetime.date or datetime.datetime
-            start date
-        end_date : datetime.date or datetime.datetime
-            end date
-
-        Returns
-        -------
-        dates : list
-            list of datetimes
-        """
-        file_list = []
-        delta_all = enddate - startdate
-        timestamps = []
-
-        for i in range(delta_all.days + 1):
-            timestamp = startdate + timedelta(days=i)
-
-            files = self._search_files(
-                timestamp, custom_templ=self.day_search_str)
-
-            file_list.extend(sorted(files))
-
-        for filename in file_list:
-            timestamps.append(self._get_orbit_start_date(filename))
-
-        timestamps = [dt for dt in timestamps if (
-                dt >= startdate and dt <= enddate)]
-        return timestamps
-
-
-class AscatL1Nc(MultiTemporalImageBase):
+class AscatL1Nc(ASCAT_MultiTemporalImageBase):
     """
     Class for reading multiple ASCAT level1 images in nc format.
 
@@ -504,69 +343,6 @@ class AscatL1Nc(MultiTemporalImageBase):
                                              exact_templ=False,
                                              ioclass_kws={
                                                  'msg_name_lookup': msg_name_lookup})
-
-    def _get_orbit_start_date(self, filename):
-        """
-        Returns the datetime of the file.
-
-        Parameters
-        ----------
-        filename : full name (including the path) of the file
-
-        Returns
-        -------
-        dates : datetime object
-            datetime from the filename
-        """
-        # if your data comes from the EUMETSAT EO Portal this function can
-        if self.eo_portal == True:
-            filename_base = os.path.basename(filename)
-            fln_spl = filename_base.split('_')[4]
-            fln_datetime = fln_spl[:-1]
-            return datetime.strptime(fln_datetime, self.datetime_format)
-
-        else:
-            orbit_start_str = os.path.basename(filename)[
-                              self.filename_datetime_format[0]:
-                              self.filename_datetime_format[1]]
-            return datetime.strptime(orbit_start_str,
-                                     self.filename_datetime_format[2])
-
-    def tstamps_for_daterange(self, startdate, enddate):
-        """
-        Returns the possible timestamps of the given daterange as a datetime
-        array.
-
-        Parameters
-        ----------
-        start_date : datetime.date or datetime.datetime
-            start date
-        end_date : datetime.date or datetime.datetime
-            end date
-
-        Returns
-        -------
-        dates : list
-            list of datetimes
-        """
-        file_list = []
-        delta_all = enddate - startdate
-        timestamps = []
-
-        for i in range(delta_all.days + 1):
-            timestamp = startdate + timedelta(days=i)
-
-            files = self._search_files(
-                timestamp, custom_templ=self.day_search_str)
-
-            file_list.extend(sorted(files))
-
-        for filename in file_list:
-            timestamps.append(self._get_orbit_start_date(filename))
-
-        timestamps = [dt for dt in timestamps if (
-                dt >= startdate and dt <= enddate)]
-        return timestamps
 
 
 def get_file_format(filename):
